Get rid of rkt specific runtime code
