+++ cmd/kube-spawn/kube-spawn.go
@@ -46,14 +46,16 @@ var (
 		},
 	}
 
-	version        string
-	k8srelease     string
-	k8sruntime     string
+	version      string
+	k8srelease   string
+	k8sruntime   string
+	printVersion bool
+	debugOutput  bool
+	kubeSpawnDir string
+
 	rktBin         string = os.Getenv("KUBE_SPAWN_RKT_BIN")
 	rktStage1Image string = os.Getenv("KUBE_SPAWN_RKT_STAGE1_IMAGE")
 	rktletBin      string = os.Getenv("KUBE_SPAWN_RKTLET_BIN")
-	printVersion   bool
-	kubeSpawnDir   string
 
 	kubeadmCgroupDriver     string
 	kubeadmRuntimeEndpoint  string
@@ -63,11 +65,18 @@ var (
 
 func init() {
 	cmdKubeSpawn.Flags().BoolVarP(&printVersion, "version", "V", false, "Output version information")
+	cmdKubeSpawn.PersistentFlags().BoolVar(&debugOutput, "debug", false, "Print logs with log.Llongfile")
 	cmdKubeSpawn.PersistentFlags().StringVarP(&k8sruntime, "container-runtime", "r", defaultRuntime, "Runtime to use for the spawned cluster (docker or rkt)")
 	cmdKubeSpawn.PersistentFlags().StringVarP(&k8srelease, "kubernetes-version", "k", k8sStableVersion, "Kubernetes version to spawn, \"\" or \"dev\" for self-building upstream K8s.")
 	cmdKubeSpawn.PersistentFlags().StringVarP(&kubeSpawnDir, "kube-spawn-dir", "d", kubeSpawnDirDefault, "path to kube-spawn asset directory")
 
 	cmdKubeSpawn.PersistentPreRun = func(cmd *cobra.Command, args []string) {
+		if debugOutput {
+			log.SetFlags(log.Llongfile)
+		} else {
+			log.SetFlags(0)
+		}
+
 		// TODO: we should eventually run extensive env/config checks prior to running any subcommand
 		// That would also benefit from some kind of centralized config
 		var err error
